backend/requirements.txt
fastapi
uvicorn
firebase-admin
pandas
openpyxl
python-dotenv
